Clean up rust map backend selection
