+++ examples/rust-map/Cargo.toml
@@ -6,22 +6,34 @@ license.workspace = true
 publish = false
 repository.workspace = true
 
+[features]
+default = ["vulkan"]
+metal = [
+  "dep:objc2",
+  "dep:objc2-app-kit",
+  "dep:objc2-foundation",
+  "dep:objc2-metal",
+  "dep:objc2-quartz-core",
+]
+opengl = ["dep:glow", "dep:glutin", "dep:glutin-winit", "dep:windows-sys"]
+vulkan = ["dep:ash", "dep:ash-window"]
+
 [dependencies]
-ash = "0.38.0"
-ash-window = "0.13.0"
-glow = "0.16.0"
-glutin = "0.32.3"
-glutin-winit = "0.5.0"
+ash = { version = "0.38.0", optional = true }
+ash-window = { version = "0.13.0", optional = true }
+glow = { version = "0.16.0", optional = true }
+glutin = { version = "0.32.3", optional = true }
+glutin-winit = { version = "0.5.0", optional = true }
 maplibre-native = { path = "../../bindings/rust/crates/maplibre-native" }
 raw-window-handle = "0.6.2"
 winit = { version = "0.30.13", features = ["rwh_06"] }
 
 [target.'cfg(target_os = "macos")'.dependencies]
-objc2 = "0.5.2"
-objc2-app-kit = { version = "0.2.2", features = ["NSResponder", "NSView", "objc2-quartz-core"] }
-objc2-foundation = { version = "0.2.2", features = ["NSError", "NSGeometry", "NSObject", "NSString"] }
-objc2-metal = { version = "0.2.2", features = ["MTLCommandBuffer", "MTLCommandEncoder", "MTLCommandQueue", "MTLDevice", "MTLDrawable", "MTLLibrary", "MTLPixelFormat", "MTLRenderCommandEncoder", "MTLRenderPass", "MTLRenderPipeline", "MTLResource", "MTLTexture"] }
-objc2-quartz-core = { version = "0.2.2", features = ["CALayer", "CAMetalLayer", "objc2-metal"] }
+objc2 = { version = "0.5.2", optional = true }
+objc2-app-kit = { version = "0.2.2", features = ["NSResponder", "NSView", "objc2-quartz-core"], optional = true }
+objc2-foundation = { version = "0.2.2", features = ["NSError", "NSGeometry", "NSObject", "NSString"], optional = true }
+objc2-metal = { version = "0.2.2", features = ["MTLCommandBuffer", "MTLCommandEncoder", "MTLCommandQueue", "MTLDevice", "MTLDrawable", "MTLLibrary", "MTLPixelFormat", "MTLRenderCommandEncoder", "MTLRenderPass", "MTLRenderPipeline", "MTLResource", "MTLTexture"], optional = true }
+objc2-quartz-core = { version = "0.2.2", features = ["CALayer", "CAMetalLayer", "objc2-metal"], optional = true }
 
 [target.'cfg(target_os = "windows")'.dependencies]
-windows-sys = { version = "0.61.2", features = ["Win32_Foundation", "Win32_Graphics_Gdi"] }
+windows-sys = { version = "0.61.2", features = ["Win32_Foundation", "Win32_Graphics_Gdi"], optional = true }
