+++ examples/rust-map/mise.toml
@@ -1,25 +1,47 @@
 [tasks.compile-shaders]
 hide = true
-run = [
-  "glslangValidator -V src/vulkan_texture_compositor/shaders/fullscreen.vert -o src/vulkan_texture_compositor/shaders/fullscreen.vert.spv",
-  "glslangValidator -V src/vulkan_texture_compositor/shaders/sample.frag -o src/vulkan_texture_compositor/shaders/sample.frag.spv",
-]
+run = '''
+set -euo pipefail
+
+if [[ "${MLN_FFI_RENDER_BACKEND:-}" == "vulkan" ]]; then
+  glslangValidator -V src/vulkan_texture_compositor/shaders/fullscreen.vert -o src/vulkan_texture_compositor/shaders/fullscreen.vert.spv
+  glslangValidator -V src/vulkan_texture_compositor/shaders/sample.frag -o src/vulkan_texture_compositor/shaders/sample.frag.spv
+fi
+'''
 
 [tasks.build]
 depends = ["//:ensure-native-library", ":compile-shaders"]
 dir = "{{env.MLN_FFI_REPO_ROOT}}"
-run = "cargo build -p rust-map"
+run = '''
+case "${MLN_FFI_RENDER_BACKEND:-}" in
+  metal | opengl | vulkan) feature="$MLN_FFI_RENDER_BACKEND" ;;
+  *) echo "unsupported rust-map render backend: ${MLN_FFI_RENDER_BACKEND:-unset}" >&2; exit 1 ;;
+esac
+cargo build -p rust-map --no-default-features --features "$feature"
+'''
 
 [tasks.ci]
 depends = [":build"]
 dir = "{{env.MLN_FFI_REPO_ROOT}}"
-run = "cargo clippy -p rust-map --all-targets -- -D warnings"
+run = '''
+case "${MLN_FFI_RENDER_BACKEND:-}" in
+  metal | opengl | vulkan) feature="$MLN_FFI_RENDER_BACKEND" ;;
+  *) echo "unsupported rust-map render backend: ${MLN_FFI_RENDER_BACKEND:-unset}" >&2; exit 1 ;;
+esac
+cargo clippy -p rust-map --no-default-features --features "$feature" --all-targets -- -D warnings
+'''
 
 [tasks.run]
 depends = ["//:ensure-native-library", ":compile-shaders"]
 dir = "{{env.MLN_FFI_REPO_ROOT}}"
 usage = 'arg "render-target"'
-run = 'cargo run -p rust-map -- "{{ usage.render_target }}"'
+run = '''
+case "${MLN_FFI_RENDER_BACKEND:-}" in
+  metal | opengl | vulkan) feature="$MLN_FFI_RENDER_BACKEND" ;;
+  *) echo "unsupported rust-map render backend: ${MLN_FFI_RENDER_BACKEND:-unset}" >&2; exit 1 ;;
+esac
+cargo run -p rust-map --no-default-features --features "$feature" -- "{{ usage.render_target }}"
+'''
 
 [tasks."run:owned-texture"]
 run = [{ task = "run", args = ["owned-texture"] }]
